+++ packages/style-value-types/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 Style Value Types adheres to [Semantic Versioning](http://semver.org/).
 
+## [3.1.6] 2019-07-25
+
+### Fixed
+
+- Clamping color alpha to 0-1.
+
 ## [3.1.4] 2019-05-01
 
 ### Fixed